localtest/requirements.txt
sympy
mpmath
Pillow
numpy
sklearn
matplotlib